typo: Missing parenthesis in copy/paste example
In the PCBNew plugin documentation, there is a copy paste example to
present the present the end user with all discovered plugin paths on the
system.

Previously the example was using a Python 2 style syntax of `print`
without parenthesis.  This commit adds the parenthesis as allowed by the
`future` import and Python 3.x